# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/wizard/generate:
    post:
      operationId: generatePageDraft
      tags: [wizard]
      summary: Generate landing-page copy from a brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WizardGenerateInput"
      responses:
        "200":
          description: Generated page draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WizardResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/templates:
    get:
      operationId: listTemplates
      tags: [wizard]
      summary: List saved landing-page templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PageTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTemplate
      tags: [wizard]
      summary: Save a reusable landing-page template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateInput"
      responses:
        "200":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/templates/{id}:
    delete:
      operationId: deleteTemplate
      tags: [wizard]
      summary: Delete a template
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/affiliate/links:
    get:
      operationId: listAffiliateLinks
      tags: [growth]
      summary: List affiliate/partner links with click counts
      responses:
        "200":
          description: Affiliate links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AffiliateLink"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAffiliateLink
      tags: [growth]
      summary: Create an affiliate/partner link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateLinkInput"
      responses:
        "200":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateLink"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/affiliate/links/{id}:
    put:
      operationId: updateAffiliateLink
      tags: [growth]
      summary: Update an affiliate/partner link and its keywords
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateLinkInput"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateLink"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAffiliateLink
      tags: [growth]
      summary: Delete an affiliate/partner link
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/affiliate/suggest:
    post:
      operationId: suggestAffiliateLinks
      tags: [growth]
      summary: AI-suggest partner-link placements in a page's draft copy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestLinksInput"
      responses:
        "200":
          description: Suggested placements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkSuggestionsResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/affiliate/apply:
    post:
      operationId: applyAffiliateLinks
      tags: [growth]
      summary: Insert selected partner links into a page's draft copy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyLinksInput"
      responses:
        "200":
          description: Updated page draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      description: Numeric resource identifier.
      schema:
        type: integer
...
  schemas:
...
    Locale:
      type: string
      enum: [en, nl]
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OkResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    AdminSection:
      type: object
      description: A content section within a draft. `data` holds the localized content for its `type`.
      properties:
        type:
          type: string
        order:
          type: integer
        data:
          type: object
          additionalProperties: true
      required:
        - type
        - order
        - data
...
    AdminPage:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        hasUnpublishedChanges:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/AdminSection"
        updatedAt:
          type: string
      required:
        - id
        - slug
        - title
        - locale
        - status
        - hasUnpublishedChanges
        - sections
        - updatedAt
...
    WizardGenerateInput:
      type: object
      properties:
        topic:
          type: string
          minLength: 1
        persona:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        tone:
          type: ["string", "null"]
        locale:
          $ref: "#/components/schemas/Locale"
        sections:
          type: array
          items:
            type: string
        templateId:
          type: ["integer", "null"]
      required:
        - topic
        - locale
    WizardResult:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        sections:
          type: array
          items:
            $ref: "#/components/schemas/AdminSection"
      required:
        - title
        - slug
        - sections
    PageTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - name
        - config
        - createdAt
    CreateTemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        config:
          type: object
          additionalProperties: true
      required:
        - name
        - config
    AffiliateKeyword:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
        active:
          type: boolean
      required:
        - id
        - keyword
        - locale
        - active
    AffiliateKeywordInput:
      type: object
      properties:
        keyword:
          type: string
          minLength: 1
        locale:
          $ref: "#/components/schemas/Locale"
        active:
          type: boolean
      required:
        - keyword
    AffiliateLink:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetUrl:
          type: string
        description:
          type: ["string", "null"]
        sponsored:
          type: boolean
        active:
          type: boolean
        clickCount:
          type: integer
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/AffiliateKeyword"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - targetUrl
        - sponsored
        - active
        - clickCount
        - keywords
        - createdAt
        - updatedAt
    AffiliateLinkInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        targetUrl:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        sponsored:
          type: boolean
        active:
          type: boolean
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/AffiliateKeywordInput"
      required:
        - name
        - targetUrl
        - keywords
    SuggestLinksInput:
      type: object
      properties:
        pageId:
          type: integer
      required:
        - pageId
    LinkSuggestion:
      type: object
      properties:
        sectionIndex:
          type: integer
        sectionType:
          type: string
        linkId:
          type: integer
        linkName:
          type: string
        keyword:
          type: string
        snippet:
          type: string
      required:
        - sectionIndex
        - sectionType
        - linkId
        - linkName
        - keyword
        - snippet
    LinkSuggestionsResult:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/LinkSuggestion"
      required:
        - suggestions
    ApplyLinkInsertion:
      type: object
      properties:
        sectionIndex:
          type: integer
        linkId:
          type: integer
        keyword:
          type: string
      required:
        - sectionIndex
        - linkId
        - keyword
    ApplyLinksInput:
      type: object
      properties:
        pageId:
          type: integer
        insertions:
          type: array
          items:
            $ref: "#/components/schemas/ApplyLinkInsertion"
      required:
        - pageId
        - insertions
...